refactor: useCarousel 기능 분리
